fix(party-service): split consent tests into their own class
detekt's LargeClass rule failed CI on PartyServiceTest once the two new
consent tests pushed it over threshold. Move them into a dedicated
PartyServiceConsentTest instead of trimming coverage.
